requirement.txt
python
wandb
mistralai
panda
rouge_score
datasets
numpy
scikit-learn
matplotlib